# API/openapi/openapi.source.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Strava API",
    "description": "Canonical OpenAPI generated from API code and augmented with Bruno request coverage.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:7075/api"
    }
  ],
  "paths": {
...
    "/account": {
      "delete": {
        "tags": [
          "User management"
        ],
        "summary": "Submit an account delete job for the current user.",
        "operationId": "DeleteAccount_Delete",
        "responses": {
          "204": {
            "description": "CORS preflight response or account delete job queued"
          },
          "401": {
            "description": "Session is missing or invalid"
          }
        }
      }
    },
...
    "/manage/users": {
      "get": {
        "tags": [
          "Admin"
        ],
        "operationId": "GetUsersStatus_Get",
        "parameters": [
          {
            "name": "x-admin-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payload of Array of AdminUserStatus",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/adminUserStatus"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No description"
          }
        }
      }
    },
    "/manage/users/{userId}": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "operationId": "RemoveUser_Delete",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "All data belonging to user has been removed"
          }
        },
        "security": [
          {
            "function_key": []
          }
        ],
        "x-bruno-request-file": "^manage/users/remove-user.bru"
      }
    },
...
    "/user": {
      "get": {
        "tags": [
          "User management"
        ],
        "summary": "Get the authenticated user's profile and Strava sync status.",
        "operationId": "GetCurrentUser_Get",
        "parameters": [
          {
            "name": "session",
            "in": "cookie",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payload of CurrentUserResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/currentUserResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "adminUserStatus": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "activityCount": {
            "type": "integer",
            "format": "int32"
          },
          "lastLoggedInAtUtc": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
      "currentUserResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "syncStatus": {
            "$ref": "#/components/schemas/stravaSyncStatus"
          },
          "foundTotals": {
            "$ref": "#/components/schemas/foundTotalsResponse"
          }
        }
      },
...
      "foundTotalsResponse": {
        "type": "object",
        "properties": {
          "summitedPeaks": {
            "type": "integer",
            "format": "int32"
          },
          "visitedPaths": {
            "type": "integer",
            "format": "int32"
          },
          "visitedAreas": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "stravaProcessedActivityCounts": {
        "type": "object",
        "properties": {
          "summitedPeaks": {
            "type": "integer",
            "format": "int32"
          },
          "visitedPaths": {
            "type": "integer",
            "format": "int32"
          },
          "visitedAreas": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "stravaSyncStatus": {
        "type": "object",
        "properties": {
          "totalActivitiesOnStrava": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "syncedActivities": {
            "type": "integer",
            "format": "int32"
          },
          "processedActivities": {
            "$ref": "#/components/schemas/stravaProcessedActivityCounts"
          },
          "updatedAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "function_key": {
        "type": "apiKey",
        "name": "code",
        "in": "query"
      }
    }
  },
...
}